// src/compass/privacy/types.ts
export enum DataCategory {
  IDENTITY = "IDENTITY",
  FINANCIAL = "FINANCIAL",
  HEALTH = "HEALTH",
  LEGAL = "LEGAL",
  IMMIGRATION = "IMMIGRATION",
  EMPLOYMENT = "EMPLOYMENT",
  FAMILY = "FAMILY",
}